requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
